update mybadge in stats page, I have moved global variable userbadges code from badges.vue to Dashboard.vue
